core: separate image preparation and starting NBD server
The two operations are currently combined, however we need different
behavior when each fails. If image preparation fails we can go straight
to endWithFailure. But if starting the NBD server failed we would have
to tear down the image.

This patch splits them apart, and starts the NBD server after the image
is prepared and the transfer entity is created. This way cleanup will be
performed and the image will be deactivated should starting the NBD
server fails.

Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>