core: do not send bitmaps for full backups
Currently we try to run incremental backup on disks that are RAW if they
have a QCOW overlay, this is bound to fail and we should use full backup
in this case.

This patch checks the backup mode of the disk and considers it when
deciding whether to start the NBD server with a bitmap.

Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>