core: respect deprecated custom property mdev_type
We dropped the custom property mdev_type from the predefined properties
in favor of (fake) VM devices. However, this breaks backward compatibility
with clients that specify the vGPU settings using this custom property.

Therefore patch changes add-VM and update-VM to respect the mdev_type
custom property when specified. Note that the custom property was already
converted to VM devices when reading an OVF and that covered most of the
cases, like snapshots that were taken while mdev_type existed - that
code is now being reused in the aforementioned flows.