core: code refactoring of validators related to VM Import
The changes in the VM import flow due to adding VMs shared disks
to be part of the OVF, required the validators update.
This commit is focused on some code refactoring prior to introducing
"real" changes in order to separate main change and code improvements.

- Better names to:
  - Make code more clear which of the validators are checking disks
and which validators check disks images.
  - Make it clear by the names that we the validators verify
that the disks do *not* exist.
- Improving a log message in case of a duplicate disk's image.
- Local variables renaming in the validation methods to improve readability.
- Some warnings, comments fixed & long lines fixed.

Signed-off-by: Pavel Bar <pbar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1922977
Change-Id: If211bbed8a9c89f82990d2fdd684819005a3c314