scheduling: consider numa memory to dedicated
When we have virtual NUMA provided with a dedicated VM, we wish to pin
that virtual NUMA to a physical NUMA. But we have a memory constraint.
The future physical resources we are going to use needs to provide
enough memory in order to use it.
This needed to be considered when scheduling such VM and selecting the
physical resources to use.
For now the NUMA will use the tune `interleave`, that would let the
memory pages to be used from multiple physical NUMA nodes.

Change-Id: I9d3a79296391cd77f7d53cc75eb152383dba9137
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>