webadmin: Disable Stateless VM field
When the latest version of the template is selected
in the new VM dialog, the VM has to remain stateless
no matter what is set on the template. To ensure that
(and to allow the selection of the latest stateful template),
the field is automatically disabled and checked as Stateless.