core: measure volume before starting conversion
When converting a disk on block storage to qcow2, we might not have
enough space for the qcow metadata on the created volume if it is
converted from RAW format.

This change measures the required size before creating the skeleton
volume, and if the required size is smaller than required, the returned
size will be used.

Bug-Url: https://bugzilla.redhat.com/2069670
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>