core: Fix removal of unmanaged network with long name
The long name networks are represented on the host as onUUID.
When the network is unmanaged and the attachment is gone we need
to remove network based on it's vdsm name and not the name
that it had in engine. Use vdsmName for removal of unmanaged
interfaces instead of the engine name.

Signed-off-by: Ales Musil <amusil@redhat.com>
Bug-Url: https://bugzilla.redhat.com/2070053