engine: keep original cpuProfile in UpdateVmVersionCommand
If the vm and the template are not on the same cluster,
adding of the newly created vm fails because of the validation
of the cpu profile (it requires the cpu profile to be from the vm's
cluster, but it is actually copied from the template on a different cluster).

This patch keeps the original cpuProfile in vm's configuration if the
template is on a different cluster.