core: fix racing cluster fips settings
When we have a new cluster, the FIPS mode is defined by the first host
being added to that cluster. By default the value of it when unavailable
from the host is `false`. In case of new host deployment to that cluster
and reboot, we may not get the capabilities as expected and even when
the host is with FIPS set, it will still wrongly set the cluster with
`false`, making the host non-operational until the cluster is edited by
the user to be FIPS `true`.

The placement the engine handles the FIPS is moved to a place we can be
certain we already get the capabilities from the host. Therefore, we
will always set it with the right value of the first host in that new
cluster.

Since the value from the host is now used on cluster level and used
internally only, this information was removed from the host general tab
in the UI.

Change-Id: I3c21028e3bd0f882340afc13ab05911fb92ec90c
Bug-Url: https://bugzilla.redhat.com/2065543
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>