core: consider exclusively pinned cpus when scoring and balancing hosts
When the host is scored, its total CPU usage is divided
by the number of cores. This works well when all CPUs
are shared because we expect the underlying systems to
distribute the load equally.

However, with dedicated CPUs, the load is not equal. We do not
really care about the load on the dedicated CPUs as we cannot run
any other VM there. We care about the remaining CPUs to see
if there is enough resources left for the VMs already running
on the host.

This patch changes the calculation of the host cpu load so that
the load is calcualted only for the shared CPUs.