core: inline HandleVdsFipsCommand into InitVdsOnUp
Now that we check the host FIPS mode (compared to the cluster's mode) in
InitVdsOnUp, we need to take special care about the VDS object we check
since InitVdsOnUp is called before persisting the data that is received
by the host monitoring. So instead of passing the reported data to
HandleVdsFips, it would be simpler to inline HandleVdsFips into
InitVdsOnUp - we have everything accessible there and we can reuse some
existing code to set the non-operational mode, if needed.

Bug-Url: https://bugzilla.redhat.com/2065543
Signed-off-by: Arik Hadas <ahadas@redhat.com>