core: reverse snapshots list
Currently, when removing individual disk snapshots, they are grouped
by their respective disk_id and are removed together, in the order from
the parent to the child. For example:
A <- S2 <- S1

And we select disks from S1 and S2, S1 will be removed first and then
S2. However, since block commit was introduced, we only do forward
merges, meaning when removing S1, the child S2, will be the one removed
after its contents are merged into S1. So when the operation attempts to
merge S2, it will no longer exist.

This patch fixes the issue by reversing the order of images to start the
operation the other way around.

Bug-Url: https://bugzilla.redhat.com/2070451
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>