core: always update volume size when creating a new image
Vdsm often aligns the size of the volume when creating, and the volume
will end up having a different size in the engine than what is on the
storage. This is already handled when a new disk is created but not when
a new snapshot is created.

Since there is already a getVolumeInfo request at the end of snapshot
creation, we can use this opportunity to update the size of the volume
as well.

Bug-Url: https://bugzilla.redhat.com/1852308
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>