core: adjust volume size when converting to cow/preallocated
When converting to COW/Prellocated from RAW additional size may be
required for the qcow header. We already measure required size, however
we only use it for the initial size of the volume which is not used when
the target is a preallocated volume on block storage.

This patch uses required size to increase the target volume if it's
larger than the current size of the volume to ensure there is enough
space to perform the conversion.

Bug-Url: https://bugzilla.redhat.com/2076047
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>