core: Further limit the maximum number of vCPUs
With the lifted limit on the maximum number of CPU sockets implemented
in commit a4c08581377d669300ee35d1cd6406f0b37f25f0, the maximum
possible number of vCPUs increased significantly.  But it is not
without its price.  The increased maximum number of vCPUs reduces
memory available to the guest OS and may cause some problems,
especially with UEFI firmware.

There is already MaxNumOfCpusCoefficient config value limiting the
maximum number of vCPUs.  But it is not enforced strictly, the maximum
number of vCPUs can still be significantly higher, to be consistent
with the previous behavior.  Let’s be more cautious and clearer and
apply MaxNumOfCpusCoefficient strictly, unless the maximum number of
vCPUs is not very large, in the sense of the newly introduced
ManyVmCpus config value.

Bug-Url: https://bugzilla.redhat.com/2075486