core: Increase TSEG size for large VMs with UEFI
Large VMs using UEFI may not start if the TSEG size is not big
enough.  There is no clear rule when the TSEG size must be increased.
We increase it in the following cases:

- If the maximum number of VM vCPUs is at least 128.

- If the maximum memory is at least 16 GB, i.e. the memory taken from
  the guest by changing the TSEG size is negligible.

We increase the TSEG size by wild-guess values based on various
comments from Bugzilla, see the comment in the code for more details.
As the TSEG size is taken from the memory available to the guest, we
do not want to make it unnecessarily large by default.  On the other
hand, it’s better to be on the safe side and allocate a bit more than
perhaps needed rather than causing the VM not to start.

We introduce config values for both the limits mentioned above, in
case the selected values don’t work well enough.

Bug-Url: https://bugzilla.redhat.com/2075486