packaging: Renew expiring vmconsole certificates
In commit 35e8f5110af4a84c69eda38a8f5f56d26b050594, a check was added
to renew vmconsole certificates when the CA certificate is newer.  But
the certificates are still not renewed if they are expired or close to
being expired.  Let’s fix it by additionally checking for certificate
expiration.

In order to that, we must check the helper certificate file instead of
key.

Bug-Url: https://bugzilla.redhat.com/1988496