core: redundant "Date" object's generation during Copy Constructor
Some "Date" data members are copied using reference copying, while in
some 3 places (without any visible reason) a new object is created.

Changing for consistency to a simple reference copy: if we want to
protect from changing an internal "Date" object via one of the
containing "DiskImage" instances, we probably need to do it in
many more places in this class (and other classes as well).

Signed-off-by: Pavel Bar <pbar@redhat.com>