webadmin: CPU pinning error message wording
When there is no host found where the VM can run,
the following message is displayed:

The host ABC did not satisfy internal filter CpuPinning because...

It makes sense to continue with "it" after that.

Also, in case of manual and resize and pin pinning, we should
mention also the possibility that some cores have been exclusively pinned.