core: Fix applying of the pending resources
In VdsCpuUnitPinningHelper::previewPinOfPendingExclusiveCpus,
the CPU topology has been updated with an incorrect
CPU pinning policy (the current VM's instead of the pending one).

This may have caused that after the migration of a VM group
with mixed exclusive and manual pinnings, some CPUs might
end up to be pinned to both, manual and exclusive CPUs at once.