api: block on execution of ImportVmFromConfiguration
we don't normally to this but in case of ImportVmFromConfiguration it makes
some sense to invoke it as a blocking (sync) operation since it doesn't involve
disk operation so if may only have async operation related to vm leases which are
relatively quick. The main reason to do this is that some backup applications
relied on getting Created response and can't cope with Accepted but it also makes
to do that because it is often the case that clients try to start the VM right
after the execution of ImportVmFromConfiguration is completed.

Bug-Url: https://bugzilla.redhat.com/2074112
Signed-off-by: Arik Hadas <ahadas@redhat.com>