certificates: Deal with host certificates expiration sooner
Currently, we warn about host certificate expiration 120 days in
advance and alert and perform renewal on host upgrades 30 days in
advance.  These proved to be too late, many users don’t pay
appropriate attention to warnings and alerts and when there is no host
upgrade during the relatively short period, the users may end up with
expired certificates and non-responsive hosts.

Now, when we have extended the lifetime of host certificates to 3650
days, we can warn about host certificate expiration earlier and force
renewal on host upgrades already during the warning period.  This
increases the chances there will be a host upgrade (with certificate
renewal) during the period.  Value of CertExpirationWarnPeriodInDays
is increased to 365.

This change also extends the warning period for all other certificates
except for web certificates.

Bug-Url: https://bugzilla.redhat.com/2079890