certificates: Extend the lifetime of non-web certificates
Non-web certificates are not subjects to the lifetime limitations
required by web browsers.  VDS certificates must be renewed when the
host is in maintenance, which makes frequent upgrades difficult.
Let’s extend the lifetime of non-web certificates to 5 years (a value
selected by Michal) to make the certificate updates much less
frequent.

Bug-Url: https://bugzilla.redhat.com/2079835