webadmin: Allow enrolling certificates in non-responsive states
When a host certificate expires the host becomes non-responsive.
There is no longer a way to connect to the host via the Vdsm API.
But it would be still possible to renew the certificate if the
operation wasn’t blocked in the web UI.  Let’s permit enrolling
certificates from the menu also for hosts in non-responsive state, to
make life easier for users who ignore warning about upcoming
certificate expiration until it actually happens.

Of course, a host can become non-responsive also for other reasons
than its certificate expiration.  If the users decide to renew
certificates in such states, it’s their own decision and risk.
A host in such a state is not in a good shape anyway and restarting
the services running on it, as needed by the certificate renewal, is
not that much an issue.

Bug-Url: https://bugzilla.redhat.com/2079901