restapi: prefer parentType over subclasses in getParent
Currently LinkHelper#getParent may return subclasses even if the
parentType exists in the methods list.

This makes the returned resource unpredicatble as the method list does
not have any guaranteed order.

To fix this, this patch tries to match parentType first, and only if it
is not in the list it will fallback to a subclass.

Bug-Url: https://bugzilla.redhat.com/2080728
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>