ansible: Restart services after Vdsm certificate update
After updating Vdsm certificates, the corresponding services are not
restarted.  Or, actually, they are, but only as a part of
ovirt-host-deploy-vnc-certificates role if those certificates happened
to be updated as well (they usually do), which is not entirely
sufficient.  And ovirt-imageio is not restarted at all.

Let’s make sure the services are always restarted after Vdsm
certificates update.  It’s sufficient to restart libvirtd (it restarts
vdsmd etc. too) and ovirt-imageio (vdsmd wants it started but doesn’t
ensure its restart).

We check for both ‘enabled’ status and ‘running’ states.  Some
services may be stopped before updates, so checking for ‘running’
wouldn’t be sufficient.  And ovirt-imageio may be running but not
enabled, which is probably a bug but it doesn’t harm to check for
both.

This change also means that some services may be restarted more than
once during a certificate update.  Fixing this would need introducing
inter-role service handling, which would be too complicated at the
moment.  Multiple restarts should be usually harmless because this is
normally done in the maintenance mode.

Bug-Url: https://bugzilla.redhat.com/2079901