webadmin: Fix reseting of graphics
When the BIOS type changed, the graphics list was always
reset even though the video type stayed the same.

This was because the setting of the items and selected item
was separated in https://gerrit.ovirt.org/c/ovirt-engine/+/118132

This patch introduces a new constructor ListModel that allows
to configure the selection model. That allows to set the video type
field with NO_SELECTION model and avoid the synchronization that
caused problems mentioned in https://gerrit.ovirt.org/c/ovirt-engine/+/118132.

Bug-Url: https://bugzilla.redhat.com/2078193