build: use maven 3.6 --no-transfer-progress
The github action CI build log tends to contain a large amount of
log lines for the transfer progress of downloading dependencies.
This can make a prohibitively large log file to be able to view
it on the github action page.

Use maven 3.6 and the option `--no-transfer-progress` to vastly reduce
the log output to where we can actually scroll the the bottom of the
output on the action's web page.

Note: maven is a separate module on centos 8 stream.  Copr chroot
configs will need to be checked.  Enabling at least maven 3.6 is
requried.