copr: update build to configure `safe.directory` as needed
See [CVE-2022-24765](https://github.blog/2022-04-12-git-security-vulnerability-announced/)

git >= 2.35.2 won't work on copr `make_srpm` builds without marking
the working directory as a `safe.directory` in git.

With copr using the `make_srpm` build SRPM type, the source build will
first create a git repo as per the package source configurations.  Then
it uses mock to runs the `.copr/Makefile srpm` target with the working
directory set to the root of the initial git clone.  Since the user
that created the git repo is different from the user running in the
mock container, the CVE mitigation is hit and `git` will fail to run.

This change explicitly adds the working directory as a `safe.directory`
allowing git commands used during the `srpm` build to function.