core: avoid null value so the ${DiskName} will be injected
if somehow withoption function get null as variablevalue it will not
try to inject variablevalue instead of variableName so this cause the message to
look like this:
Cannot remove Virtual Disk. Disk ${DiskName} is being moved or copied.
so I added validation of getParameters().getNewAlias() and getParameters().getDiskAlias()
and if both are null the LockMessage will be somthing more genaric and will look like this:
Cannot remove Virtual Disk. Related operation is currently in progress. Please try again later.
and if at list one of them is not null I can inject into the message so the message will look like this:
Cannot remove Virtual Disk. Disk vm_disk is being moved or copied.

Bug-Url: https://bugzilla.redhat.com/2039746
Signed-off-by: Artiom Divak <adivak@redhat.com>