core: Use edited VM configuration when cloning
Configuration of a cloned VM may be edited before cloning. In this case
it is passed in CloneVmParameters#vm and should be saved in editedVm
variable and passed to UpdateVmCommand that is executed after cloning.

The code that saves this configuration was incorrectly removed by change
6f143de593b2ce415e7155b8e3a1d634637afa74. This patch restores this code
with one difference: UpdateVmCommand is now executed by CoCo, so
editedVm was moved from CloneVmCommand to parameters.

Change-Id: Id77061ba6749351dcfbe832ad1c38288c6ada479
Bug-Url: https://bugzilla.redhat.com/2080758
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>