webadmin: trigger 'cancel' action on MoveOrCopyDiskModel
In case the environment has a large amount of disks, and triggering the
move/copy dialog opening and closing for a couple of times, the browser
crashes due to a memory leak caused by a huge amount of property change
listeners that remain although pressing the 'cancel' button.
By triggering the 'cancel' operation including 'cleanup()', the browser
doesn't crash as before.

Bug-Url: https://bugzilla.redhat.com/2001574
Signed-off-by: Shani Leviim <sleviim@redhat.com>