scheduling: switch CpuPinningPolicy to group vms
This patch will change the CpuPinningPolicyUnit to handle a VM group
instead of a single VM.
This change became necessary since we introduced the `dedicated` CPU
feature. Once it being called to migration of multiple VMs, one doesn't
know what the previous VM took in terms of resources. Therefore, it show
a possible migration but fail on the migration itself.
Now, the data will be saved locally to check if the host is possible for
this operation and filter it out in case it doesn't.

Change-Id: Ia74cba8162c43c99674ca3b2b6855099566b0ff9
Bug-Url: https://bugzilla.redhat.com/2079351
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>