core: override storage domain id when creating target disk
When copying from a traditional domain to an MBS domain we create a new
ManagedBlockStorageDisk that is based on the source disk with updated
paramteters.

Currently createManagedBlockDiskFromDiskImage adds the target
Storage Domain to the target's disk list, however, this list already
contains the source domain, which makes the disk present on two Storage
Domains.

This patch overrides the original disks Storage Domain list.