scheduling: reduce CPU pinning fragmentation
This patch sorts the VM list on schedule phase before doing the
dedicated CPU pinning.
This change tries to reduce the fragmentation that may be caused when
allocating a smaller VM in terms of the policy and CPU topology (a less
restricted) when having a bigger one. Scheduling the smaller VM first
may cause less physical resources needed by the larger VM because we try
to enhance performance in the allocation logic.
Now, the VMs will be sorted based on the policy first and if needed by
the CPU topology, letting the more restricted VM to be scheduled first.

Change-Id: Ia3fc28511c147d43da9cf8d86d50ca17f1e3f0e1
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>