persist event notifier service boots
This bug is a result of applying level 3 security of the new USM
introduced on SNMP V3.
When the Event Notification Service is restarts , it creates a new USM
object but does not update the number of time its was rebooted.
This makes the service suspected by SNMP V3 and it stops getting traps
from it in snmptrapd.
The solution is to add a new configuration value
NotificationServiceBoots which is set by default to 0.
Each time the Event Notification Service is rebooted, this value is
incremented by 1 and stored in the configuration table (vdc_options)

When a new USM object is created on Event Notification Service run
method, it set the number of boots accordint to the persisted value.

This solves the problem and Event Notification Service can be restated
several times and still keep getting traps on which it is listening
according to its "FILTER" settings.

Signed-off-by: Eli Mesika <emesika@redhat.com>
Bug-Url: https://bugzilla.redhat.com/2072626