db: fix fn_db_change_column_null
Changing the function that sets a column in the part that turns from
NULL to not NULL to set any found NULL values on the changed column
values to a default value according to the column type.

This script changes also upgrade scripts to pass the v_type new
parameter.

In case that unsupported type is passed, an error is thrown.

This patch also removed redundant code added in PRs 297 and 300

Signed-off-by: Eli Mesika <emesika@redhat.com>
Bug-Url: https://bugzilla.redhat.com/2077387
Allow-db-upgrade-script-changes: Yes