Block snapshot removal when VM in backup
Currently removing snapshots on VM during Hybrid backup is not blocked
which allows user to remove the parent snapshot of the auto-generated
snapshot. As a result, the user cannot initiate an image transfer.

Blocking any snapshot removal during active backup fixes this issue,
and also prevents other conflicts. This validation is skipped for
snapshot removal done as part of Hybrid Backup finalization