webadmin: block editing when resize and pin
Before this patch, the edit VM failed for
all VMs with Resize and Pin NUMA CPU Pinning Policy
if the memory, cpu or NUMA was modified.

Now it is possible to edit memory for VMs
with Resize and Pin NUMA without any restrictions.
Editing of CPU and NUMA is still not allowed, but instead
of the backend validation failure, the fields are reset to their
original values when Resize and Pin NUMA is selected.

Bug-Url: https://bugzilla.redhat.com/2074522