core: allow setting vnuma nodes without pinning
This patch will allow setting via UI and API virtual NUMA nodes to a VM
without having the VM pinned to a host. If a `numa_node_pin` is set, the
scheduler will filter out hosts that won't fit to the NUMA pinning.
Otherwise, the VM will start with the set amount of virtual NUMA nodes.

Change-Id: Ia35a95babf3ae6b2d4713c809052be1d6ddede38
Bug-Url: https://bugzilla.redhat.com/2081410
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>