core: set limit the unknown status in ansible
When calling ansible callback, we check the event and see the command
status. Initially the command status is `unknown`. Until now the async
ansible callback had no limitations, but if the ansible runner service
and the artifacts are gone, it will keep being on `unknown` state or
move to this state again, running for unlimited time.
A new configuration, `AsyncAnsibleTimeout` added, by default set to 180
seconds as a time limit.
When we reach the time limit we will try to cancel the playbook in order
to release the usage of the relevant entities.

Change-Id: I31c7c4a1bc17721806c70769377fe1a1cf059b57
Bug-Url: https://bugzilla.redhat.com/2030293
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>