don't throw an exception when canceling playbook if play execution was already completed
In some cases play execution is completed, but host deploy process is still running and parsing events.
In such case if the timeout was reached, an attempt to cancel the playbook will fail because the play process
no longer exists.
In cases where the command to cancel the playbook fails, an error will be logged with both host-deploy and
stdout logs.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=2087738