fencing: adding port option to ipmilan agent
The fencing agents has special options and custom options.
The special options includes: "secure", "port" and "slot"
The custom options includes any other option
Up to now there was no need to enable "port" setting in ipmilan
In order to enable "port" as an optional field there is a new setting
for ipmilan agent for version 4.7 in vdc_options VdsFenceOptionMapping
option_name that adds this field such a placeholder for the "port" field
is created in the UI.
This new "port" field is not mandatory and existing hosts with ipmilan
agents omitting the "port" value will work as before that change.

Signed-off-by: Eli Mesika <emesika@redhat.com>
Bug-Url: https://bugzilla.redhat.com/2087886