engine: Change the memory usage calculation
So far, the memory usage was calculated in VDSM and
the calculation considered hugepages as used memory. As a result,
the host with hugepages had high memory usage even though the
hugepages were not used.

This patch moves the calculation from VDSM to the engine. This
allows to calculate the memory usage differently in various
scenarios.

a) In the UI, the memory usage calculation includes hugepages so that
the user has the full overview of the host memory
b) In host monitoring, we issue VDS_HIGH_MEM_USE only for the normal
memory and calculation does NOT include hugepages

Bug-Url: https://bugzilla.redhat.com/2006625