core: vnic hot un\plug - avoid race
Lack of a lock in ActivateDeactivateVmNicCommand allows several
instances of hot plug and unplug requests to run in parallel causing
inconsistency in vNIC status as well as libvirt failing the requests due
to non-unique aliases. This patch adds a lock to the command so that
each instance of the command has to finish before the next one starts.

Bug-Url: https://bugzilla.redhat.com/2084530
Change-Id: I93ac9f1f12f34526ad10b9a49790c3a2689e8c4d
Signed-off-by: Eitan Raviv <eraviv@redhat.com>