core: don't use vdsm core for isolate threads
When allocating CPUs for isolate threads we use the maximum freed
socket. We get all the non-pinned CPUs for this socket. The problem is
that later on we don't filter out cores with pinned CPUs. Now we return
all the CPUs in the socket and that way, when we iterate the cores, we
are able to filter out those with pinned CPUs.

Change-Id: Ic81d93ecd830cea689a8eea4579d6e728c5b2e04
Bug-Url: https://bugzilla.redhat.com/2094415
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>