core: Make sure at most one ramfb device is enabled
When a VM has more than one ramfb vGPU devices enabled, it fails to
start.  ramfb is used only during boot, before the guest OS loads its
GPU driver (or if the GPU driver doesn’t work properly).  Then there
should be no noticeable difference for the user whether ramfb is
enabled on a given vGPU or not.

Under these circumstances, the simplest way to avoid the problem with
multiple ramfb-enabled vGPUs is to enable ramfb only for the first one
written to the domain XML, which is what this patch implements.

See the attached bug for more details.

Bug-Url: https://bugzilla.redhat.com/2079760