engine: Fix the calculation of the VM's CPU cores
After the refactoring of CpuPolicyUnit, we started
to calculate required shared CPUs with countThreadsAsCores=true.
This caused that the VMs with RESIZE_AND_PIN cpu pinning policy
were not able to migrate between identical hosts when cluster's
countThreadsAsCores=false.

This patch changes the calculation in CpuPolicyUnit back to
countThreadsAsCores=false and changes that also for the pending resources
used in CpuPolicyUnit and the minRequiredSharedCpusCount of the host.

Bug-Url: https://bugzilla.redhat.com/2095259